recap

+++ src/CSM.TmpeSync.LaneArrows/Network/Handlers/LaneArrowAppliedHandler.cs
@@ -40,7 +40,7 @@ namespace CSM.TmpeSync.Network.Handlers
                     resolvedLaneId,
                     resolvedSegmentId,
                     resolvedLaneIndex);
-                if (TmpeBridge.ApplyLaneArrows(resolvedLaneId, arrows))
+                if (TmpeBridge.ApplyLaneArrows(resolvedLaneId, (int)arrows))
                     Log.Info(LogCategory.Synchronization, "Applied remote lane arrows | laneId={0} segmentId={1} laneIndex={2} arrows={3}", resolvedLaneId, resolvedSegmentId, resolvedLaneIndex, arrows);
                 else
                     Log.Error(LogCategory.Synchronization, "Failed to apply remote lane arrows | laneId={0} segmentId={1} laneIndex={2} arrows={3}", resolvedLaneId, resolvedSegmentId, resolvedLaneIndex, arrows);
